% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{aCAPf1f2}
\alias{aCAPf1f2}
\title{Pairwise warping using the optimal combined distance}
\usage{
aCAPf1f2(
  func1,
  func2,
  alpha = 0.5,
  plots = TRUE,
  verbose = FALSE,
  n_digit = 5,
  max_drv = 10
)
}
\arguments{
\item{func1}{A matrix for curve 1 with time (equidistant) in the first column and function values in the second column.}

\item{func2}{A matrix for curve 2 with time (equidistant) in the first column and function values in the second column.}

\item{alpha}{A vector of alpha values. Default = 0.5.}

\item{plots}{Whether to plot the warping results. Default = TRUE.}

\item{verbose}{Whether to output the procedure. Default = FALSE.}

\item{n_digit}{Number of decimal points in the output. Default = 5.}

\item{max_drv}{An integer of the maximum derivative allowed for warping function. Default = 10. A smaller value decreases computation time.}
}
\value{
A list of warping results for each alpha.
\itemize{
  \item \code{warping_func}: A matrix for warping function with time in the first column and function values in the second column.
  \item \code{warped_func}: A matrix for warped function with time in the first column and function values in the second column.
  \item \code{d_A}: Amplitude shape distance.
  \item \code{d_P}: Phase distance.
  \item \code{d_C}: Optimal combined distance.
  \item \code{alpha}: The alpha values used in warping.
}
}
\description{
Pairwise warping using the optimal combined distance
}
\examples{
t = seq(0, 1, length.out = 100+1) # time
fdata1 = cbind(t, 4*exp(-(t-0.3)^2 / 0.01) + 6*exp(-(t-0.7)^2 / 0.03)) # curve 1
fdata2 = cbind(t, 7*exp(-(t-0.4)^2 / 0.03) + 3*exp(-(t-0.8)^2 / 0.01)) # curve 2
fit = aCAPf1f2(func1 = fdata1, func2 = fdata2, alpha = 0.5, verbose = FALSE)
}
